pyproject.toml
[project]
name = "website"
version = "0.1.0"
description = "Fitness trainer web app for tracking food intake, calories, and exercise"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "flask>=3.0.0",
    "gunicorn>=21.0.0",
    "authlib>=1.3.0",
    "requests-oauthlib>=1.3.0",
    "python-dotenv>=1.0.0",
    "flask-sqlalchemy>=3.1.0",
    "flask-wtf>=1.2.0",
    "wtforms>=3.1.0",
]
